requirements.txt
requests
yfinance
pandas
matplotlib
python-dotenv
streamlit
python-pptx